requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
config==0.4.0
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
dataclasses==0.6
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-dynamic-scraper==0.13.2
django-heroku==0.3.1
django-model-utils==3.1.2
future==0.17.1
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jsonpath-rw==1.4.0
lxml==4.3.3
oauthlib==3.0.1
parsel==1.5.1
Pillow==5.4.1
ply==3.11
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-decouple==3.1
python-scrapyd-api==2.1.2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.0
scrapyd==1.2.0
service-identity==18.1.0
six==1.12.0
social-auth-core==3.1.0
Twisted==18.9.0
urllib3==1.24.1
userprofile==0.2.5
w3lib==1.20.0
whitenoise==4.0
zope.interface==4.6.0
